Pull request #2852: Add 3 Scatter layers with PyMIL tests activated
Merge in COREOS/coreml from eng/PR-65676196 to eng/PR-66791560

Squashed commit of the following:

commit 5f1761e72a07566643881a4d58138d0f6fa2f0ec
Author: Gitesh Dawer <gdawer@apple.com>
Date:   Tue Aug 11 10:02:05 2020 -0700

    Address review comments

    Radar-Id: rdar://problem/65676196

commit cd797476b9d3f231714699d0162480c10ef46f4d
Author: Gitesh Dawer <gdawer@apple.com>
Date:   Mon Aug 10 10:16:33 2020 -0700

    Add 3 Scatter layers with PyMIL tests activated

    Radar-Id: rdar://problem/65676196

    Add c++, validator and tf tests

    Radar-Id: rdar://problem/65676196

+++ coremltools/converters/mil/frontend/tensorflow/test/test_ops.py
@@ -3151,12 +3151,12 @@ class TestScatter:
         self, use_cpu_only, backend, data_rank, indices_rank
     ):
 
-        shape = np.random.randint(low=2, high=4, size=data_rank)
+        shape = np.random.randint(low=2, high=4, size=data_rank).astype(np.int32)
         indices_shape = np.random.randint(low=2, high=4, size=indices_rank)
         indices_shape[-1] = np.random.randint(low=1, high=data_rank + 1)
         updates_shape = list(indices_shape[:-1]) + list(shape[indices_shape[-1] :])
 
-        updates = np.random.rand(*updates_shape)
+        updates = np.random.rand(*updates_shape).astype(np.float32)
         indices_list = []
         for i in range(indices_shape[-1]):
             indices_list.append(np.random.randint(0, shape[i], size=indices_shape[:-1]))
